update unit tests for new component
* created new area for fixture data which is used in cardigan and common tests components.
* create snapshots
