+++ cardigan/stories/content.js
@@ -1,5 +1,5 @@
 import faker from 'faker';
-
+import { openingTimes as openingTimesFixture } from '@weco/common/test/fixtures/components/opening-times';
 export function randomNumber(min, max) {
   return Math.floor(Math.random() * max) + min;
 }
@@ -285,94 +285,4 @@ export function organisation() {
   };
 }
 
-export const openingTimes = {
-  collectionOpeningTimes: {
-    placesOpeningHours: [
-      {
-        id: 'Wsttgx8AAJeSNmJ4',
-        order: 1,
-        name: 'Galleries',
-        openingHours: {
-          regular: [
-            { dayOfWeek: 'Monday', opens: null, closes: null },
-            { dayOfWeek: 'Tuesday', opens: '10:00', closes: '18:00' },
-            { dayOfWeek: 'Wednesday', opens: '10:00', closes: '18:00' },
-            { dayOfWeek: 'Thursday', opens: '10:00', closes: '21:00' },
-            { dayOfWeek: 'Friday', opens: '10:00', closes: '18:00' },
-            { dayOfWeek: 'Saturday', opens: '10:00', closes: '18:00' },
-            { dayOfWeek: 'Sunday', opens: '10:00', closes: '18:00' },
-          ],
-          exceptional: [],
-        },
-      },
-      {
-        id: 'WsuS_R8AACS1Nwlx',
-        order: 2,
-        name: 'Library',
-        openingHours: {
-          regular: [
-            { dayOfWeek: 'Monday', opens: '10:00', closes: '18:00' },
-            { dayOfWeek: 'Tuesday', opens: '10:00', closes: '18:00' },
-            { dayOfWeek: 'Wednesday', opens: '10:00', closes: '18:00' },
-            { dayOfWeek: 'Thursday', opens: '10:00', closes: '20:00' },
-            { dayOfWeek: 'Friday', opens: '10:00', closes: '18:00' },
-            { dayOfWeek: 'Saturday', opens: '10:00', closes: '16:00' },
-            { dayOfWeek: 'Sunday', opens: null, closes: null },
-          ],
-          exceptional: [],
-        },
-      },
-      {
-        id: 'WsuYER8AAOG_NyBA',
-        order: 3,
-        name: 'Restaurant',
-        openingHours: {
-          regular: [
-            { dayOfWeek: 'Monday', opens: null, closes: null },
-            { dayOfWeek: 'Tuesday', opens: '11:00', closes: '18:00' },
-            { dayOfWeek: 'Wednesday', opens: '11:00', closes: '18:00' },
-            { dayOfWeek: 'Thursday', opens: '11:00', closes: '21:00' },
-            { dayOfWeek: 'Friday', opens: '11:00', closes: '18:00' },
-            { dayOfWeek: 'Saturday', opens: '11:00', closes: '18:00' },
-            { dayOfWeek: 'Sunday', opens: '11:00', closes: '18:00' },
-          ],
-          exceptional: [],
-        },
-      },
-      {
-        id: 'WsuZKh8AAOG_NyUo',
-        order: 4,
-        name: 'Café',
-        openingHours: {
-          regular: [
-            { dayOfWeek: 'Monday', opens: '08:30', closes: '18:00' },
-            { dayOfWeek: 'Tuesday', opens: '08:30', closes: '18:00' },
-            { dayOfWeek: 'Wednesday', opens: '08:30', closes: '18:00' },
-            { dayOfWeek: 'Thursday', opens: '08:30', closes: '21:00' },
-            { dayOfWeek: 'Friday', opens: '08:30', closes: '18:00' },
-            { dayOfWeek: 'Saturday', opens: '09:30', closes: '18:00' },
-            { dayOfWeek: 'Sunday', opens: '10:00', closes: '18:00' },
-          ],
-          exceptional: [],
-        },
-      },
-      {
-        id: 'WsuaIB8AAH-yNylo',
-        order: 5,
-        name: 'Shop',
-        openingHours: {
-          regular: [
-            { dayOfWeek: 'Monday', opens: '09:00', closes: '18:00' },
-            { dayOfWeek: 'Tuesday', opens: '09:00', closes: '18:00' },
-            { dayOfWeek: 'Wednesday', opens: '09:00', closes: '18:00' },
-            { dayOfWeek: 'Thursday', opens: '09:00', closes: '21:00' },
-            { dayOfWeek: 'Friday', opens: '09:00', closes: '18:00' },
-            { dayOfWeek: 'Saturday', opens: '10:00', closes: '18:00' },
-            { dayOfWeek: 'Sunday', opens: '10:00', closes: '18:00' },
-          ],
-          exceptional: [],
-        },
-      },
-    ],
-  },
-};
+export const openingTimes = openingTimesFixture;
